add transfer service. mvp is redy now

+++ src/main/java/com/revolut/assignment/datamodel/Account.java
@@ -36,13 +36,13 @@ public class Account {
             if (newValue.compareTo(BigDecimal.ZERO) == -1) {
                 String comment = "Not Enough Money";
                 historyList.add(new History(delta, comment, false));
-                long historyIndex = historyList.size() - 1;
+                int historyIndex = historyList.size() - 1;
                 message = new Message(false, comment, historyIndex);
             } else {
                 String comment = "Success";
                 value = newValue;
                 historyList.add(new History(delta, comment, true));
-                long historyIndex = historyList.size() - 1;
+                int historyIndex = historyList.size() - 1;
                 message = new Message(true, comment, historyIndex);
             }
 
@@ -50,7 +50,7 @@ public class Account {
             String comment = "Internal Error";
             e.printStackTrace();
             historyList.add(new History(delta, comment, false));
-            long historyIndex = historyList.size() - 1;
+            int historyIndex = historyList.size() - 1;
             message = new Message(false, comment, historyIndex);
         } finally {
             semaphore.release();
